chore(medium): remove types from matrix definitions

+++ ciphers/medium.py
@@ -25,7 +25,7 @@ class Medium:
     _sbox: np.ndarray[int] = np.array([0xc, 0xa, 0xd, 0x3, 0xe, 0xb, 0xf, 0x7, 0x8, 0x9, 0x1, 0x5, 0x0, 0x2, 0x4, 0x6])
     _pbox: np.ndarray[int] = np.array([0, 2, 3, 1])
     _pbox_inverse: np.ndarray[int] = np.array([0, 3, 1, 2])
-    _m = np.array([[0, 1], [1, 0]], dtype=int)
+    _m = np.array([[0, 1], [1, 0]])
     _alpha = 0b_1001_0101_0010_1110
 
     def encrypt(self, plaintext: int, key: int, use_related_key: bool = False) -> int:
